+++ GUI_20212202_IJA9WQ/Models/LawnMover.cs
@@ -11,14 +11,14 @@ namespace GUI_20212202_IJA9WQ.Models
 {
     public class LawnMover : Moveable
     {
-        public LawnMover(int placeX, int placeY, int displayWidth, int displayHeight, float speed)
+        public LawnMover(double placeX, double placeY, double displayWidth, double displayHeight, float speed)
         {
             this.placeX = placeX;
             this.placeY = placeY;
             this.displayWidth = displayWidth;
             this.displayHeight = displayHeight;
             this.speed = speed;
-            this.actualX = placeX;
+            this.actualX = (float)placeX;
         }
 
         public override Geometry Area
